arm: Use 64-bit addressing in a bunch of places.

+++ src/common/common_types.h
@@ -47,8 +47,8 @@ typedef double f64; ///< 64-bit floating point
 
 // TODO: It would be nice to eventually replace these with strong types that prevent accidental
 // conversion between each other.
-typedef u32 VAddr; ///< Represents a pointer in the userspace virtual address space.
-typedef u32 PAddr; ///< Represents a pointer in the ARM11 physical address space.
+typedef u64 VAddr; ///< Represents a pointer in the userspace virtual address space.
+typedef u64 PAddr; ///< Represents a pointer in the ARM11 physical address space.
 
 // An inheritable class to disallow the copy constructor and operator= functions
 class NonCopyable {
